feat: Prefer package.json#scripts over rw cli
 ## Context

Now that we have a `reloaded` package, we need to decide between
* (a) having a cli wrap all the cli running
* (b) having projects just define package.json#scripts for what needs to be run

We've decided on (b):
* inevitably users need to add, wrap around or change their projects build scripts, so encapsulating it all is unlikely anyways
* with a cli, we'd be hiding what's actually happening from the user
* with a cli, we'd be maximizing our surface area, arguably unneccesarily
* we can generate this when scaffolding out the
  project

 ## Changes
* This PR changes us from approach (a) to (b)

 ### `rw` -> `rw-scripts`
* We now use `pnpm rw-scripts` for scripts that are in fact needing to live in `reloaded`, for example:
  * We don't need a script for `vite build`, that can just be in package.json#scripts as `"build": "vite build"`
  * But we _do_ need a script for `pnpm worker-run` (how users can define scripts run in the worker, e.g. seed scripts) - so we'll keep that around as a script in reloaded, and allow running it via `rw-scripts worker-run`
* The name `rw-scripts` indicates its a collection of scripts, rather than a full fledged cli
