+++ experiments/billable/package.json
@@ -8,6 +8,21 @@
   "author": "",
   "license": "UNLICENSED",
   "private": true,
+  "scripts": {
+    "build": "vite build",
+    "migrate:dev": "wrangler d1 migrations apply DB --local",
+    "migrate:prd": "wrangler d1 migrations apply DB --remote",
+    "dev": "while true; do NODE_ENV=development vite dev; [ $? -eq 0 ] || break; done",
+    "dev:init": "rw-scripts dev-init",
+    "seed": "pnpm worker:run ./src/scripts/seed.ts",
+    "migrate:new": "rw-scripts migrate-new",
+    "worker:run": "rw-scripts worker-run",
+    "clean": "pnpm build && pnpm clean:vendor",
+    "clean:vite": "rm -rf ./node_modules/.vite",
+    "clean:vendor": "rm -rf ./vendor/dist",
+    "release": "pnpm build && wrangler deploy",
+    "format": "prettier --write ./src"
+  },
   "dependencies": {
     "@cloudflare/workers-types": "^4.20241112.0",
     "@prisma/adapter-d1": "^6.3.0",
