+++ reloaded/src/scripts/build-vendor-bundles.mts
